add utf8変換

+++ ARoQEngine/GameUtility/Base/Private/Base/Include/GUStringBase.hpp
@@ -1141,7 +1141,7 @@ namespace gu::details::string
 		{ 
 			Ensure(0); 
 		}
-		return num;
+		return static_cast<float>(num);
 	}
 
 	template<class Char, int CharByte>
